Since Drucker is a single application doing several things, this PR is turning all .py files - called modules - into a collective package called "app". This means that the code under app/, should no longer be called directly from the shell or main drucker.sh script (this is why the shebang is gone).
Python packages, are basically APIs that other developers can use - which is also why a module may NEVER run code on itself when you call it from the shell or "import app.foo", in any case, nothing should happen. This is why all functions here move under a new function "main()", which is similar to the Python convention of "if __name__ == '__main__'" in main scripts.