Say hi to the new Drucker main script!
I have deleted drucker.sh and replaced it with a symlink, so other apps remain to work - don't worry about losing it (git lets you go back in time). The new command behaves 99% the same so nobody is going to notice, after we've fixed the bugs I caused of course.