Refactored Drucker from lose scripts into a single app (#56)
* Turned app/ into a real Python package instead of a directory with loose scripts, the imports also are much more readable now because in a package, you must use "from . import ..." notation. The file app/__init__.py plays a crucial role in a Python package and must be maintained, I kept it simple.

* Since Drucker is a single application doing several things, this PR is turning all .py files - called modules - into a collective package called "app". This means that the code under app/, should no longer be called directly from the shell or main drucker.sh script (this is why the shebang is gone).

Python packages, are basically APIs that other developers can use - which is also why a module may NEVER run code on itself when you call it from the shell or "import app.foo", in any case, nothing should happen. This is why all functions here move under a new function "main()", which is similar to the Python convention of "if __name__ == '__main__'" in main scripts.

* Proper shell applications, always set the correct UNIX status code so that Drucker's commands can be integrated into other bash shell scripts or CI-code. Here I'm moving 1/0 as constants into variables.py, to make things more readable throughout the code and less confusing (as 0 is good, 1 is bad).

* In line with previous commits, no modules under the app/ package should
run code on their own. In drucker.sh, you're wrapping around
arguments.py from a bash script - but as we're going to replace that,
this removed piece of logic moves into the new upcoming main file.

* For readability, I'm renaming 'mainfunc' into 'dispatched_function' to make crystal-clear that this is the main function running during Drucker runtime.

* I'm renaming the 'args' argument introduced in the last PR into 'drucker', since its role evolves a little bit. Instead of just holding the CLI-arguments, the new 'drucker' object becomes the "backbone object of Drucker", like the root of a tree. This 'drucker' object, also also has 'drucker.vars' which references to your variables, so while code can continue to 'from . import variables', you will now also be able to start using 'drucker.vars.APP_ROOT' without even importing it.

* Send the "drucker" backbone object deeper into Drucker and send it to all functions that currently reference variables using 'v.FOO'. The assertions don't do anything besides silencing Pylint, which instead will show TODO's so you don't can continue this in a different PR.

* Making PEP8/pylint a happier puppy before it barks.

* Say hi to your new best friend: exceptions.

Exceptions are "errors", but they are NOT necessarily "bugs". Error conditions - e.g. ansible check - are perfectly fit for "raising exceptions". Your code can "raise <ExceptionType>" from anywhere but Python itself can raise them to, for instance SyntaxError or KeyboardInterrupt (both types or exceptions).

Exceptions "travel back up" in your code until something catches them and does something with them.  The program can stop entirely or could continue to run using a workaround, this all depends on where you "catch" a certain type of exception. In this commit I have replaced a few error conditions where you did a sys.exit(), instead sending RuntimeError's (red message) or RuntimeWarning's (yellow) will do the same and looks much cleaner.

The next commit is the main file and the try/catch is in it.

* Say hi to the new Drucker main script!

I have deleted drucker.sh and replaced it with a symlink, so other apps remain to work - don't worry about losing it (git lets you go back in time). The new command behaves 99% the same so nobody is going to notice, after we've fixed the bugs I caused of course.

* First small bugfix: provisioning always succeeds unless Exceptions we're thrown.

* No need for drucker.sh to stay

* PEP8

* bugfix #2

* fix #3

* Improving exception output a little bit.

* Remove drucker.sh from other locations where referenced